Move Go version as the default tool.
