+++ python/MANIFEST.in
