+++ python/oauth2l/testdata/fake_service_account.json
