+++ sgauth/example/library/main.go
