+++ sgauth/internal/const.go
