+++ sgauth/settings.go
