+++ util/tasks.go
@@ -19,7 +19,7 @@ import (
 	"net/http"
 	"io/ioutil"
 	"errors"
-	"github.com/google/oauth2l/go/sgauth"
+	"github.com/google/oauth2l/sgauth"
 	"context"
 	"encoding/json"
 	"log"
